Project: Recipe book
Create a web page with your favorite recipes! The recipe book should have a table of contents linking to each recipe below, using internal links. Each recipe should have a table for the ingredients, a list of details, and a list and/or paragraphs of the steps required. If you use recipes from the internet, be sure to add a link to the website.

Use CSS to style the recipe book, using a mix of the simple selectors and colors that you've learned.

Tip: Can't find the images you want in our image selector? You can also bring in images from Wikipedia. To do that, use a site-restricted Google search, then click "View image" on the result you like, and copy the URL into your <img> tag.